+++ src/main/java/com/akiban/server/types3/common/BigDecimalWrapper.java
@@ -25,13 +25,15 @@
  */
 package com.akiban.server.types3.common;
 
+import java.math.BigDecimal;
+
 public interface BigDecimalWrapper {
     
      BigDecimalWrapper add(BigDecimalWrapper augend);
      BigDecimalWrapper subtract(BigDecimalWrapper augend);
      BigDecimalWrapper multiply(BigDecimalWrapper augend);
      BigDecimalWrapper divide(BigDecimalWrapper augend);
-     long ceil();
+     BigDecimal ceil();
      BigDecimalWrapper divide(BigDecimalWrapper augend, int scale);
      int getSign();
      void reset();
